+++ spec/rubocop/cop/style/else_alignment_spec.rb
@@ -84,12 +84,24 @@ describe RuboCop::Cop::Style::ElseAlignment do
     context 'with assignment' do
       context 'when alignment style is variable' do
         context 'and end is aligned with variable' do
-          it 'accepts an if with end aligned with setter' do
+          it 'accepts an if-else with end aligned with setter' do
             inspect_source(cop,
                            ['foo.bar = if baz',
                             '  derp1',
                             'else',
                             '  derp2',
+                            'end'])
+            expect(cop.offenses).to be_empty
+          end
+
+          it 'accepts an if-elsif-else with end aligned with setter' do
+            inspect_source(cop,
+                           ['foo.bar = if baz',
+                            '  derp1',
+                            'elsif meh',
+                            '  derp2',
+                            'else',
+                            '  derp3',
                             'end'])
             expect(cop.offenses).to be_empty
           end
@@ -134,14 +146,17 @@ describe RuboCop::Cop::Style::ElseAlignment do
         end
 
         context 'and end is aligned with keyword' do
-          it 'registers an offense for an if with setter' do
+          it 'registers offenses for an if with setter' do
             inspect_source(cop,
                            ['foo.bar = if baz',
                             '            derp1',
-                            '          else',
+                            '          elsif meh',
                             '            derp2',
+                            '          else',
+                            '            derp3',
                             '          end'])
-            expect(cop.messages).to eq(['Align `else` with `foo.bar`.'])
+            expect(cop.messages).to eq(['Align `elsif` with `foo.bar`.',
+                                        'Align `else` with `foo.bar`.'])
           end
 
           it 'registers an offense for an if with element assignment' do